refactor(api): split webhook transport god-file (signature/replay/dispatch)
`transport/webhook/transport.rs` (799 lines) mixed routing+signature policy
+replay-window+dispatch pipeline into one file. Split into focused modules:

- `signature.rs` — ADR-0022 enforcement: `enforce_signature`, metric recording
  (`record_signature_failure`), problem+json response builders, constant-time
  `PROBLEM_JSON_CONTENT_TYPE`. Imports `replay::replay_reason_for` for the
  double-bump on `NEBULA_WEBHOOK_REPLAY_REJECTIONS_TOTAL`.
- `replay.rs` — replay-window reason mapping: `replay_reason_for` translates
  signature failure reason codes to dedicated replay-counter label set.
- `dispatch.rs` — ADR-0049 single-pipe: `webhook_handler`,
  `slug_webhook_handler`, `dispatch_inner` (routing lookup -> rate-limit ->
  signature -> oneshot dispatch), `http_response_to_axum`,
  `record_rate_limit_rejection`.
- `transport.rs` — `WebhookTransport` public struct, `WebhookTransportConfig`,
  `ActivationHandle`, `TransportInner`, `ActivationError`, `generate_nonce`,
  lifecycle API. Delegates into `dispatch` for the handler fns.

Public API and behavior byte-identical. 236 tests pass (1 skipped).
